Various improvements

+++ primitives/shared/src/main/scala/com/malliina/json/PrimitiveFormats.scala
@@ -3,11 +3,11 @@ package com.malliina.json
 import play.api.libs.json.Json.toJson
 import play.api.libs.json._
 
-import scala.concurrent.duration.{Duration, DurationLong}
+import scala.concurrent.duration.{Duration, DurationDouble}
 
 object PrimitiveFormats {
   implicit val durationFormat: Format[Duration] = Format[Duration](
-    Reads(_.validate[Long].map(_.seconds)),
+    Reads(_.validate[Double].map(_.seconds)),
     Writes(d => toJson(d.toSeconds))
   )
 }
